moved particle renderer to procedural library, general cleanup
